Upgrade result saving setup processes
Saving directory info is only saved as "save-dir-name" until the "moment of save". (no set-current-directory command until then.)

User can  click the set-save-dir button to specify the dir (=update the save-dir-name String) or when "Reset" is clicked, it will check if any files needs to be saved, and if so it will guide the use to set save directory.

Case handling tested for all cases, but more rigorous testing may be needed.

Some other changes:
- Layout rearrangement.
- run-index monitor changed visible name to "Current run"
- Remove "force-save-video" button.
- Remove "stable timestep max" monitor.
- Relocate the function "initialize-saving" in the code.
- Relocate the function "initialize-saving" in the code.
- "smaller-part-test?" option condition change : smaller-part-test? and Calculation-Type = "determinist" and "geometry-setup = "None" then, do that 10 percent setup.

- Remove test comment in line 2.